fix calling method using GET request

+++ www/application/controllers/PerubahanKuliahManage.php
@@ -64,31 +64,36 @@ class PerubahanKuliahManage extends CI_Controller {
     public function answer() {
         date_default_timezone_set("Asia/Jakarta");
         try {
-            $userInfo = $this->Auth_model->getUserInfo();
-            $this->db->where('id', $this->input->post('id'));
-            $this->db->update('PerubahanKuliah', array(
-                'answer' => $this->input->post('answer'),
-                'answeredByEmail' => $userInfo['email'],
-                'answeredDateTime' => strftime('%Y-%m-%d %H:%M:%S'),
-                'answeredMessage' => htmlspecialchars($this->input->post('answeredMessage'))
-            ));
+            if ($this->input->server('REQUEST_METHOD') == 'POST'){
+                $userInfo = $this->Auth_model->getUserInfo();
+                $this->db->where('id', $this->input->post('id'));
+                $this->db->update('PerubahanKuliah', array(
+                    'answer' => $this->input->post('answer'),
+                    'answeredByEmail' => $userInfo['email'],
+                    'answeredDateTime' => strftime('%Y-%m-%d %H:%M:%S'),
+                    'answeredMessage' => htmlspecialchars($this->input->post('answeredMessage'))
+                ));
 
-            $this->load->model('Email_model');
-            $query = $this->db->get_where('PerubahanKuliah', ['id' => $this->input->post('id')]);
-            $row = $query->row();
-            $subject = "Status Permohonan Perubahan Kuliah Anda";
-            $message = $this->load->view('PerubahanKuliahManage/email', array(
-                'answeredByName' => $this->bluetape->getName($userInfo['email']),
-                'requestByName' => $this->bluetape->getName($row->requestByEmail),
-                'id' => $row->id,
-                'mataKuliahCode' => $row->mataKuliahCode,
-                'mataKuliahName' => $row->mataKuliahName,
-                'answer' => $row->answer,
-                'answeredMessage' => $row->answeredMessage
-            ), TRUE);
-            $this->Email_model->send_email($row->requestByEmail, $subject, $message);
+                $this->load->model('Email_model');
+                $query = $this->db->get_where('PerubahanKuliah', ['id' => $this->input->post('id')]);
+                $row = $query->row();
+                $subject = "Status Permohonan Perubahan Kuliah Anda";
+                $message = $this->load->view('PerubahanKuliahManage/email', array(
+                    'answeredByName' => $this->bluetape->getName($userInfo['email']),
+                    'requestByName' => $this->bluetape->getName($row->requestByEmail),
+                    'id' => $row->id,
+                    'mataKuliahCode' => $row->mataKuliahCode,
+                    'mataKuliahName' => $row->mataKuliahName,
+                    'answer' => $row->answer,
+                    'answeredMessage' => $row->answeredMessage
+                ), TRUE);
+                $this->Email_model->send_email($row->requestByEmail, $subject, $message);
+                
+                $this->session->set_flashdata('info', 'Permintaan perubahan kuliah sudah dijawab dan email notifikasi sudah dikirimkan.');
+            } else {
+                throw new Exception("Can't call method from GET request!");
+            }
             
-            $this->session->set_flashdata('info', 'Permintaan perubahan kuliah sudah dijawab dan email notifikasi sudah dikirimkan.');
         } catch (Exception $e) {
             $this->session->set_flashdata('error', $e->getMessage());
         }
@@ -97,10 +102,15 @@ class PerubahanKuliahManage extends CI_Controller {
 
     public function remove() {
         try {
-            $id = $this->input->post('id');
-            $this->db->where('id', $id);
-            $this->db->delete('PerubahanKuliah');
-            $this->session->set_flashdata('info', "Permohonan #$id telah dihapus.");
+            if ($this->input->server('REQUEST_METHOD') == 'POST'){
+                $id = $this->input->post('id');
+                $this->db->where('id', $id);
+                $this->db->delete('PerubahanKuliah');
+                $this->session->set_flashdata('info', "Permohonan #$id telah dihapus.");
+            } else {
+                throw new Exception("Can't call method from GET request!");
+            }
+            
         } catch (Exception $e) {
             $this->session->set_flashdata('error', $e->getMessage());
         }
